+++ src/main.ts
@@ -2,6 +2,7 @@ import path from "path"
 import express from "express"
 import dotenv from "dotenv"
 
+import { RequestRange } from "./types"
 import { parseStatement } from "./parseStatement"
 import { handleStatement } from "./handleStatement"
 
@@ -13,8 +14,10 @@ const statementPath = path.resolve(process.env.STATEMENT_PATH || "")
 const app = express()
 
 app.get("/**", async (req, res) => {
+    const { from, to } = req.query as RequestRange
+
     const statement = await parseStatement(statementPath)
-    const report = handleStatement(statement)
+    const report = handleStatement(statement, { from, to })
     res.json(report)
 })