+++ src/parseStatement.ts
@@ -2,6 +2,7 @@ import { createReadStream, PathLike } from "fs"
 import { parse } from "csv-parse"
 
 import { Transaction } from "./types"
+import { getTimestampByDate } from "./utils"
 
 export async function parseStatement(filePath: PathLike) {
     const promise = (): Promise<Transaction[]> =>
@@ -13,6 +14,7 @@ export async function parseStatement(filePath: PathLike) {
                 .on("data", (csvRow) => {
                     const transaction: Transaction = {
                         date: csvRow[0],
+                        timestamp: getTimestampByDate(csvRow[0]),
                         ticker: csvRow[1] || null,
                         type: csvRow[2],
                         quantity: csvRow[3] ? Number(csvRow[3]) : null,