+++ src/utils.ts
@@ -2,7 +2,9 @@ export function getTimestampByDate(dateTime: string) {
     const [date, time] = dateTime.split(" ")
     const [DD, MM, YYYY] = date.split("/")
 
-    return Number(new Date(`${YYYY}-${MM}-${DD} ${time}`))
+    const isoDate = `${YYYY}-${MM}-${DD}`
+
+    return Number(new Date(time ? `${isoDate} ${time}` : isoDate))
 }
 
 export function roundAmount(number: number) {